+++ packages/service/core/workflow/dispatch/child/runApp.ts
@@ -21,7 +21,7 @@ import { ReadPermissionVal } from '@fastgpt/global/support/permission/constant';
 import { getAppVersionById } from '../../../app/version/controller';
 import { parseUrlToFileType } from '@fastgpt/global/common/file/tools';
 import { getUserChatInfoAndAuthTeamPoints } from '../../../../support/permission/auth/team';
-import { getRunningUserInfoByTmbId } from '../../../../support/permission/user/team';
+import { getRunningUserInfoByTmbId } from '../../../../support/user/team/utils';
 
 type Props = ModuleDispatchProps<{
   [NodeInputKeyEnum.userChatInput]: string;
